+++ addon/objects/uml-primitives/fd-uml-object-in-state.js
@@ -2,12 +2,14 @@
   @module ember-flexberry-designer
 */
 
+import joint from 'npm:jointjs';
+
 import { computed } from '@ember/object';
 import { isArray } from '@ember/array';
-import { ActiveState } from './fd-uml-active-state';
 import { isBlank } from '@ember/utils';
+
+import { BaseObject } from './fd-uml-baseobject';
 import FdUmlElement from './fd-uml-element';
-import joint from 'npm:jointjs';
 
 /**
   An object that describes an `Object in state` element on the UML diagram.
@@ -68,11 +70,19 @@ export default FdUmlElement.extend({
 
   @for FdUmlObjectInState
   @class ObjectInState
-  @extends flexberry.uml.ActiveState
+  @extends flexberry.uml.BaseObject
   @namespace flexberry.uml
   @constructor
 */
-export let ObjectInState = ActiveState.define('flexberry.uml.ObjectInState', {
+export let ObjectInState = BaseObject.define('flexberry.uml.ObjectInState', {
+  minHeight: 40,
+  minWidth: 40,
+}, {
+  getRectangles() {
+    return [
+      { type: 'header', element: this },
+    ];
+  },
 });
 
 joint.shapes.flexberry.uml.ObjectInStateView = joint.shapes.flexberry.uml.ActiveStateView.extend({